Добавляет тестовые данные, описывает типы сущностей, реализует приложение с CLI интерфейсом, добавляет shebang. Корректировка
