+++ src/cli.ts
@@ -1,4 +1,4 @@
-#!
+#!/usr/bin/env node
 import VersionCommand from './cli-command/version-command.js';
 import HelpCommand from './cli-command/help-command.js';
 import ImportCommand from './cli-command/import-command.js';
