+++ tsconfig.json
@@ -28,8 +28,7 @@
     ]
   },
   "include": [
-    "src/**/*.ts"
-, "src/type/offer.js"  ],
+    "src/**/*.ts" ],
   "exclude": [
     "node_modules",
     "**/*.test.ts"
